+++ R/across.R
@@ -1,0 +1,34 @@
+#' Apply a function across a selection of columns
+#'
+#' @description
+#' Apply a function across a selection of columns. For use in `arrange.()`,
+#' `mutate.()`, and `summarize.()`.
+#'
+#' @param .cols vector `c()` of unquoted column names. `tidyselect` compatible.
+#' @param .fns Functions to pass. Can pass a list of functions.
+#' @param ... Other arguments for the passed function
+#' @param .names A glue specification that helps with renaming output columns.
+#' `{.col}` stands for the selected column, and `{.fn}` stands for the name of the function being applied.
+#' The default (`NULL`) is equivalent to `"{.col}"` for a single function case and `"{.col}_{.fn}"`
+#' when a list is used for `.fns`.
+#'
+#' @export
+#'
+#' @examples
+#' test_df <- data.table(
+#'   x = rep(1, 3),
+#'   y = rep(2, 3),
+#'   z = c("a", "a", "b")
+#' )
+#'
+#' test_df %>%
+#'   mutate.(across.(c(x, y), ~ .x * 2))
+#'
+#' test_df %>%
+#'   summarize.(across.(c(x, y), ~ mean(.x, na.rm = TRUE)), .by = z)
+#'
+#' test_df %>%
+#'   arrange.(across.(c(y, z)))
+across. <- function(.cols = everything(), .fns = NULL, ..., .names = NULL) {
+  abort("across.() can only work inside of tidytable verbs")
+}