+++ tests/testthat/test-summarize_across.R
@@ -1,6 +1,3 @@
-defaultW <- getOption("warn")
-options(warn = -1)
-
 test_that("single function works", {
   test_df <- tidytable(a = 1:3, b = 4:6, z = c("a", "a", "b"))
 
@@ -129,7 +126,6 @@ test_that("can pass list of named functions with .by and .names", {
 })
 
 test_that("can pass list of named functions with .by and .names using fn and col", {
-
   # This test will need to be removed when {col} and {fn} is deprecated
 
   test_df <- tidytable(a = 1:3, b = 4:6, z = c("a", "a", "b"))
@@ -143,5 +139,3 @@ test_that("can pass list of named functions with .by and .names using fn and col
   expect_equal(result_df$max_a, c(2, 3))
   expect_equal(result_df$max_b, c(5, 6))
 })
-
-options(warn = defaultW)