+++ NEWS.md
@@ -1,11 +1,13 @@
 # tidytable 0.6.1 (in development)
 
-#### Functionality improvements
-* `pivot_longer.()`: Can now use `names_to = c(".value", NA)` to drop the id column.
-
 #### New functions
 * `across.()`
 * `coalesce.()`
+
+#### Functionality improvements
+* `group_split.()`: Added a `.named` argument. If `.named = TRUE`, the resulting list will
+have names that indicate the group they are a part of.
+* `pivot_longer.()`: Can now use `names_to = c(".value", NA)` to drop the id column.
 
 #### Bug fixes
 * `if_all.()`/`if_any.()`